Log draining of broker by bosh as a maintenance state
Currently maintenance window for service fabrik gets updated
as part of deploy scripts and since docker and broker are
part of same bosh release the deploy of bosh are considered as
SF udpate. This does not help in identifying explicitly when
broker is getting updated. To address this issue, a drain script
is added which logs the status to maintenance info (if configured)
The value of this state can be checked by scheduler before
shutting itself down.